rainbow_proj/requirements.txt
django-startcbv==0.2.3
